+++ src/domain/ports/subscription.repository.ts
@@ -1,5 +1,5 @@
 import { Frequency, Subscription } from '../entities/subscription.entity.js';
-import { CreateSubscriptionDto } from '../dtos/create-subcription.dto.js';
+import { CreateSubscriptionDto } from '../dtos/create-subscription.dto.js';
 
 export interface SubscriptionRepository {
   createSubscription(dto: CreateSubscriptionDto): Promise<Subscription>;
